// frontend/types.ts
export interface Reward {
  name: string;
  rarity: 'common' | 'uncommon' | 'rare' | 'legendary';
  image: string;
}

/**
 * Extends the base Reward with properties specific to being inside a case.
 */
export interface CaseReward extends Reward {
  id: string;
  chance: number; // The drop chance percentage (e.g., 5 for 5%).
}

/**
 * Defines the structure of a reward case that users can open.
 */
export interface Case {
  id: string;
  name: string;
  price: string; // Promo code required for all cases
  image: string;
  rewards: CaseReward[]; // An array of possible rewards.
}